T.L added quickfix to bug
